backend/requirements.txt
# OSPF Demo Backend Dependencies
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0

# Redis
redis>=5.0.0

# AI - Gemini (new SDK)
google-genai>=1.0.0

# HTTP client
httpx>=0.26.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.0

# SSE for real-time updates
sse-starlette>=1.6.0

# Email validation
email-validator>=2.1.0
